Rewrite GradeSelect for grade moderation using InstUI Select
Closes FOO-2060
flag=none

This had been temporarily fixed to use a superclassed <SimpleSelect>
in order to get it to work, but that's not a good long-term solution as
it relies on <SimpleSelect> being a class component in InstUI, and it
will likely get turned into a function component at some point. The real
answer is to rewrite <GradeSelect> entirely to use the main <Select>
component.

I also had to make browser confirm dialogs asynchronous in order to
address a Chrome bug (that I'm not sure how I triggered but it seems
to be pretty clear), and as a result of that had to adjust several
specs to expect those confirms to appear asynchronously. I also had
to skip several Selenium specs that have become flakey. I was unable
to reproduce the flakiness locally and unable to figure out what
might be going on in the time box I set for myself.

Test plan:
* Enable moderated grading for your account
* Have a course with multiple teachers / graders
* In a course, create an assignment and enable moderated grading
* Have each of the teachers submit a grade for every student in the
  assignment
* Then as the "final grader" go to the grade moderation page, where
  you should see a list of each student and the provisional grade
  that each of the graders gave.
* The selection on the right should allow you to pick the score from
  one of the graders, or you should be able to type in a totally
  different score (a "custom score") and that should get added to the
  list of available scores and can be selected
* This should provide all the same functionality as the previous
  component and should work almost entirely the same way (there may
  be tiny differences due to the rewrite)
* A full set of Jest tests was written for this component using the
  same list of spec descriptions from the old component, so the
  functionality is quite well tested in the automation

Change-Id: I9fe21b72f8c92ae39a4ba2717182df146f6b3a08
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/283270
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Sean Scally <sean.scally@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>