Fail a test on route changes used by dev keys
If a developer key is given a scope (meaning that it only works
on some API endpoints), and then the route for one of those endpoints
is changed, this causes a problem for the developer key. So, we want
to prevent that. This will force that change to be more intentional.

flag=none
refs INTEROP-7476

Change-Id: I4bbf2509112f375354cdde56a6e3ab46b77f239e
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/283706
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Evan Battaglia <ebattaglia@instructure.com>
QA-Review: Evan Battaglia <ebattaglia@instructure.com>
Product-Review: Tucker Mcknight <tmcknight@instructure.com>