fix off-month mini-calendar day color contrast
The date's color is defined by the css variable
--ic-brand-font-color-dark-lightened-30
which is the result of the --ic-brand-font-color-dark color
being lightened by 30%. This does not provide the necessary
contrast. This commit replaces that var with
--ic-brand-font-color-dark-lightened-28. By lightening the color
just a bit less, we get a contrast of 5:1.
This will update any place where the css var was used.

closes LS-3076
flag=none

test plan:
  - go to the calendar
  > expect the date numbers for days not in the current month in
    the main calendar and the mini-calendar have adequate
    contrast > 4.5:1
  - go to a course syllabus
  > expect the mini calendar there to be good too

Change-Id: If16ce385d7a84a1ef7802a0efc310c5cc62a0fef
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/288938
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>