Update Caliper Docs for Unknown User Case
We recently made a change in our Live Events project to fix an issue
where events that were clearly initiated by a user were having their
actor type miscategorized as SoftwareApplication. This was due to Canvas
sometimes not including a user id, but still including a request id. For
these new cases, we now correctly categorize the event as having an
actor of type Person, but because we don't know the actual user id,
we default to 0. This commit updates our Caliper docs to reflect
this change.

In addition, it looks like when we migrated all of our Live Events docs
to the Canvas API docs, we forgot to link to the Caliper docs on the
sidebar. This commit also fixes that so that users can actually find our
all of our documentation on Caliper.

refs INTEROP-6218

flag=none

test-plan:
* Pull these changes down
* Visit the Caliper documentation
* See the change reflected

Change-Id: Ieca67d4ac46e8aa579eedb73b4cc7f482a8d2a48
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/289424
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>