Fix a11y Alert issues with Dev Key UI
Previously, when a developer key was saved at all, any alert that popped
up after closing the modal was not read by screenreaders. This was due
to the InstUI Modal component returning focus back to the element that
opened it, in order to follow a11y best practices for keyboard users.
However, this focus change effectively "steamrolled" the alert, causing
the screenreader to think that the user had tabbed away as soon as the
alert popped up. This meant that the screenreader ended up just reading
the description of the edit developer key button.

This commit fixes the issue, albeit in a less than ideal way. After a
lot of tweaking and even component rewriting, it was determined that the
easiest way (possibly only way) to avoid this issue was to use setTimeout
to ensure the alert gets rendered after the modal changes focus.

In addition, this commit also adds an aria-live attribute to the spinner
that renders while saving a key so that screenreader users know their
key is actually saving.

flag=none

fixes INTEROP-7398

test-plan:
- You'll need VoiceOver to test this, so Linux users beware.
- Enable VO. All interactions with elements should be done with your
  keyboard to really test this properly.
- Go to the developer key UI for an account and start editing an
  existing key.
- Click the save button in the modal.
- You should hear 'Saving key' read out
- Once the modal closes, you should hear 'Save successful' read out,
  followed by a description of the edit developer key button.
- Repeat the steps above but this time creating a new developer key. VO
  should read out the same things about saving the key and saving the
  key successfully.

Change-Id: I8fac3d23621a3131ccf7fce93f6a98874cb2c644
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/290103
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>