Convert `CopyToClipboardButton` to TypeScript
To allow for easier consumption within the Setting UI, this commit
converts `CopyToClipboardButton` to TypeScript. It also fixes an error
that would occur when the button was clicked at an insecure origin
(because `navigator.clipboard` is not available in that scenario). Now,
the component just renders nothing if the clipboard API is unavailable.

refs DE-1136
flag=none

test plan:
- confirm that tests pass
- confirm that the copy to clipboard button still works correctly
- confirm that at insecure origins, the button is not rendered

Change-Id: I9c292acebf5e2eacaba40860f7439c822c0d6835
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/290718
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Isaac Moore <isaac.moore@instructure.com>