Fix id-label assocation in account settings
Fixes FOO-2830
flag=none

Looks like the label value for the broken checkbox was set
incorrectly (should be :value as for all the other checkboxes
in this raft of settings, but it was set to the same thing as
the fields_for label). This resulted in the "label for=" not
matching the id of the checkbox it's meant to be associated to.

Test plan:
* see the steps to reproduce in the ticket
* the text description of that checkbox should be read out
  by the screenreader
* another hint is that the entire item (checkbox and text label)
  should get a screenreader highlight box, not just the checkbox
  itself
* NOTE you must have "Restrict students from accessing courses
  before start date" checked in order for the checkbox being
  discussed here will appear on that page!

Change-Id: Ia7052eca43d39cd9216e87cf9dcdb11764c0ca39
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/291506
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jacob Burroughs <jburroughs@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>