Revert "Page Views table I18n Date Fix and React redesign Pt. 2"
This reverts commit 5ac9047fd1bd1b55df65e49feeae77aaa89b82f1.

Reason for revert: This commit causes both a bug when reading a null
userAgent string, whole pageview box breaks, and also removes a hover
feature that was previously there. The original reason for this ticket
was to fix a translation bug, but ballooned to a larger scope.

CLOSES OREO-1647

Change-Id: I1dc2d138125b094d6825d2c93fc358cba3fd2fa7
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/291597
Reviewed-by: Bryant Larsen <blarsen@instructure.com>
QA-Review: Kwok Lam <klam@instructure.com>
Product-Review: Noah Rush <noah.rush@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>