Update minimum browser versions
flag=none

This file controls the "you are on an unsupported browser"
alert message and should be updated to reflect the most
recent version minus one of our supported browsers. In
this case I used the "long term support" version of
Firefox which is 91, along with version 94 of Chrome
(because that's what our own seleniums test against in
Jenkins) and 15 of Safari.

Test plan:
* you agree with the browser versions listed here

Change-Id: I1f2c026e689f72c9e4493538abebdc5a32e1ec23
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/292135
Reviewed-by: Ben Rinaca <brinaca@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>