Fix inconsistent plagiarism report permissions
Previously, due to permissions pecularities, if the plagiarism report
visibility was set to "never", teachers would not be able to view
plagiarism reports after their enrollments were "concluded" (due to them
not having the "manage_grades" permission then). This was inconsistent,
however, since the score would still show on speedgrader (but not the
submission page) but clicking it would error when trying to view the
report. Also, preventing them from seeing reports is especially
nonsensical since the teachers are usually the ones to set up the
assignment plagiarism report visibility which prevented them from
viewing them.

This commit makes the permissions for on the submissions page, and
actually viewing the report, nearly the same as that for using
speedgrader, although is difficult to get perfect. The permissions for
speedgrader are:
  - to use speed grader at all:
    manage_grades OR view_all_grades
  - additionally, the turnitin score is only shown if
    manage_grades OR (@context.concluded? && can_do(@context, @current_user, :read_as_admin))
    (see app/views/gradebooks/speed_grader.html.erb:235)

This commit brings the permissions for showing the score on the
submissions page, or viewing the report, to:
  - manage_grades OR view_all_grades

This means it's still possible for some cases for the score / report to
be visible from the submission page, but not from speed grader -- if:
  - the user can view_all_grades
  - the user cannot manage_grades (in the ticket this was just because
    the course/teacher enrollment was concluded)
  - the course is not concluded OR the user cannot read_as_admin

This seems like it should be very rare and I'm not sure it's worth
trying to tweak the permissions in speedgrader (which honestly I don't
understand well -- there may be some other reason for limiting what is
shown in speedgrader) or making the conditions for viewing the report as
complex as those in speedgrader. And, the risk is only that the score
will not show in speedgrader but will in the submission, not that there
will be any broken link.

fixes INTEROP-7312
flag=none

Test plan:
- setup to repro the original problem. Here is how I did this:
  - set up plagiarism platform in an assignment with report visibility
    "never" (students can never view reports). I used the
    lti_originality_report_example tool.
  - as a student, submit to the assignment
  - create an originality score/report for the submission. I used tok
    (Set up to get a token with the lti_2_token endpoint):
      tok web.canvas-lms2.docker/api/lti/assignments/469/submissions/1428/originality_report \
        originality_report[originality_score]=33 \
        originality_report[workflow_state]=scored \
        originality_report[originality_report_url]=http://example.com/123
  - make the course concluded
      course.update workflow_state: :completed
    that wil make all the enrollments also "completed".
- Make sure the original problem can now be repro'd: In this scenario
  (with the code before this commit), when viewed as a teacher, the
  submission page (e.g. courses/123/assignments/456/submissions/789,
  where 789 is the user id) should not show a plagiarism score;
  speedgrader will show a score, but when clicked, no report will be
  viewable.
  NOTE: it seems like I had to run "AdheresToPolicy::Cache.clear" and/or
  "Rails.cache.clear" for permissions to get updated when switching code
  for some reason.
- with this code checked out, make sure that:
  - the score shows up on speedgrader
  - clicking it works to show the report
  - the score shows up on the submission page and clicking it works to
    show the report
  - as a student, the score should NOT show up on the submission page

Change-Id: I6d527884205bb48de6dda98f9eac96939834f2f5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/293531
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Tucker Mcknight <tmcknight@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>