Remove Dev Key Account Binding index endpoint
While in the progress of ensuring that Developer Keys within Canvas work
within a CMC, we discovered that the Dev Key Account Bindings index
endpoint is entirely unused. Not only that, but it was not CMC-aware and
even had a bug in it that meant it broken in production.

Rather than fix and update this unused endpoint, we decided to rip it
out. If we ever do decide we need this endpoint, we can easily undo this
commit, fix the cross-shard bug, and make it CMC aware.

flag=none

refs INTEROP-7406

test-plan:
- the specs all pass
- you can no longer send a request to
  GET accounts/:account_id/developer_key_account_bindings
- if you want to be extra thorough, checkout this commit and visit the
  Dev Key UI page. If you can still view keys and their state, as well
  as update it, all is well.

Change-Id: Ifc315ee029eaa105d113702b8be5d725c19b40ad
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/294681
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Ryan Hawkins <ryan.hawkins@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>