Fix contrast issues with CondensedButton in canvas-planner
Refs FOO-3129
flag=none

The use of InstUI <CondensedButton> to replace the
deprecated <Button variant="link"> created some issues
with customers who decided to invert the contrast of
their "Primary Button" style in the theme editor. To
wit, if the Primary Button is changed from something
light-on-dark to something dark-on-light, then an instance
of a CondensedButton would have very poor contrast, as
its text is rendered in the color of the background of a
regular Button.

Test Plan:
* Have a student set up to use the Canvas Planner
* The planner should look like it used to
* Extra credit: make sure the K5 experience is also
  unchanged (should only be different in color)

Change-Id: Ia2658ba7ec67e9cca7f23cdb46058383472692b3
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/303331
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>