Clean up buttons and links in AssociationsTable
Refs FOO-3129
flag=none

In investigating this component to address the contrast
issues with <CondensedButton> I ran into a fair amount of
work... the original implementation wrapped an InstUI
component in an HTML form and then intercepted the onSubmit
action and dug around in the formdata. This seems like it's
unnecessarily complicated, so I rewrote it all to simply
pass the required data in as parms in the onClick action.

Seems to work fine ¯\_(ツ)_/¯

Test plan:
* Have a blueprint course
* While viewing it, pull out the blueprint config tray from
  the right side and click on "Associations"
* That page, including all the delete "X" and "Undo" buttons,
  should still work and look as they always have
* Pay particular attention to where focus goes as those two
  buttons are used

Change-Id: Ibe0d8d570f2486a1d2493a4a374092d39eea8a52
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/303750
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>