Fix API not returning published pace if available
The `#new` endpoint is used to either fetch the pace for the given
context or instantiate one if needed. There is a database index ensuring
there is only ever one "active" pace for any given context. However, that
database index does not ensure there are multiple unpublished paces for
any given context. Therefore, there could be a possibility where there
could be a published and unpublished pace for any given context.

This could cause the error referenced in this patchset because of the way
`#take` works. It selects the first N records without any implied order.
If a context had an unpublished and published pace when the API call for
`#new` came in it had the possibility to return the unpublished pace.
If that unpublished pace was used to try and publish it would throw an
error because we can't have two published paces for the same context.

Querying first for any published pace followed by a query for unpublished
ensures that a published pace is returned if one exists.

fixes LS-3610
flag=course_paces

test plan:
- Have a published course pace
- In the rails console, create more pace plans for that course with a
workflow state of "unpublished"
- When editing the course pace in the modal, verify the id of the object
returned is the published course pace
- Open the modal a few times to verify that the published pace is the
only one ever returned

Change-Id: I3f768f9af66f8ac35fc14bb348d365af6af04e30
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/305438
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Luis Oliveira <luis.oliveira@instructure.com>
QA-Review: Luis Oliveira <luis.oliveira@instructure.com>
Product-Review: Eric Saupe <eric.saupe@instructure.com>